+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 ## Latest changelog
 
+- Updated WaitForTaskCompletition function to handle skipping alerts
+- Updated command line argument --silent processing
+- Added remote before and after command execution hook
+- Added local before and after command execution hook
+
 14/06/2013
 ----------